# swagger.json
{"swagger":"2.0","info":{"title":"My App","description":"This is an awesome app!","version":"1.0.0"},"paths":{"\/deploy":{"post":{"summary":"GitHub webhook endpoint","description":"GitHub webhooks should fire against this endpoint.","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"type":"object"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Webhook"]}},"\/deployments\/undeploy":{"post":{"summary":"Undeploys an active deployment","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"$ref":"#\/definitions\/UndeployActionDto"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/secrets\/add":{"post":{"summary":"Adds secrets to a deployment","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"$ref":"#\/definitions\/AddSecretsActionDto"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/secrets\/purge":{"post":{"summary":"Removes all secrets of a deployment","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"$ref":"#\/definitions\/PurgeSecretsActionDto"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/deployments":{"get":{"summary":"Lists all deployments","responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/deployments\/link":{"post":{"summary":"Links a deployment","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"$ref":"#\/definitions\/LinkDeploymentActionDto"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/deployments\/unlink":{"post":{"summary":"Unlinks a deployment","consumes":["application\/json"],"produces":["application\/json"],"parameters":[{"name":"body","in":"body","required":true,"description":"Payload","schema":{"$ref":"#\/definitions\/UnlinkDeploymentActionDto"},"format":"application\/json"}],"responses":{"200":{"description":"Erfolgreich","schema":{"$ref":"#\/definitions\/ResponseDto"}}},"tags":["Endpoints"]}},"\/doc":{"get":{"responses":{"default":{"description":""}}}}},"definitions":{"ResponseDto":{"properties":{"meta":{"$ref":"#\/definitions\/MetaDto"},"data":{"type":"object"}},"type":"object"},"UndeployActionDto":{"required":["deploymentName"],"properties":{"deploymentName":{"type":"string"}},"type":"object"},"AddSecretsActionDto":{"required":["secrets","deploymentName"],"properties":{"secrets":{"items":{"items":{"type":"string"},"type":"array","maxItems":2,"minItems":2},"type":"array"},"deploymentName":{"type":"string"}},"type":"object"},"PurgeSecretsActionDto":{"required":["deploymentName"],"properties":{"deploymentName":{"type":"string"}},"type":"object"},"LinkDeploymentActionDto":{"required":["deploymentName"],"properties":{"deploymentName":{"type":"string"}},"type":"object"},"UnlinkDeploymentActionDto":{"required":["deploymentName"],"properties":{"deploymentName":{"type":"string"}},"type":"object"},"MetaDto":{"properties":{"code":{"type":"integer"},"errors":{"items":{"type":"string"},"type":"array"}},"type":"object"}},"basePath":"\/deployment\/pollux-webhook"}